+++ librapid/src/openclConfigure.cpp
@@ -230,7 +230,7 @@ __kernel void testAddition(__global const float *a, __global const float *b, __g
 		} else {
 			// Otherwise, prompt the user to select a device
 			int64_t deviceIndex = -1;
-			while (deviceIndex < 0 || deviceIndex >= global::openclDevices.size()) {
+			while (deviceIndex < 0 || deviceIndex >= int64_t(global::openclDevices.size())) {
 				std::string prompt =
 				  fmt::format("Select OpenCL device [0-{}]: ", global::openclDevices.size() - 1);
 				scn::prompt(prompt.c_str(), "{}", deviceIndex);
